hadoop_pipeline/requirements.txt
pandas>=2.2.0
scikit-learn>=1.5.0
joblib>=1.4.0
numpy>=1.26.0
nltk>=3.9
python-dateutil>=2.9.0
